Removes more usages of db-nano module
