+++ api/package.json
@@ -45,6 +45,7 @@
     "pass-stream": "^1.0.0",
     "pouchdb-adapter-http": "^7.0.0",
     "pouchdb-core": "^7.0.0",
+    "pouchdb-find": "^7.0.0",
     "pouchdb-mapreduce": "^7.0.0",
     "properties": "^1.2.1",
     "request": "^2.88.0",
