+++ api/src/db-nano.js
@@ -4,8 +4,7 @@
  *          !!! DEPRECATION WARNING !!!          *
  *************************************************/
 
-var path = require('path'),
-  url = require('url'),
+var url = require('url'),
   nano = require('nano');
 
 const { COUCH_URL, UNIT_TEST_ENV } = process.env,
@@ -16,14 +15,6 @@ if (UNIT_TEST_ENV) {
   module.exports = {
     fti: function() {},
     request: function() {},
-    getPath: function() {},
-    settings: {
-      protocol: 'http',
-      port: '123',
-      host: 'local',
-      db: 'medic',
-      ddoc: 'medic',
-    },
     use: function() {},
     medic: {
       view: function() {},
@@ -66,44 +57,6 @@ if (UNIT_TEST_ENV) {
   module.exports.medic = db.use(dbName);
   module.exports.audit = db.use(auditDbName);
   module.exports._users = db.use('_users');
-
-  module.exports.settings = {
-    protocol: parsedUrl.protocol,
-    port: parsedUrl.port,
-    host: parsedUrl.hostname,
-    db: dbName,
-    auditDb: auditDbName,
-    ddoc: 'medic',
-  };
-
-  if (parsedUrl.auth) {
-    var index = parsedUrl.auth.indexOf(':');
-    module.exports.settings.username = parsedUrl.auth.substring(0, index);
-    module.exports.settings.password = parsedUrl.auth.substring(index + 1);
-  }
-
-  module.exports.getPath = function() {
-    return path.join(
-      module.exports.settings.db,
-      '_design',
-      module.exports.settings.ddoc,
-      '_rewrite'
-    );
-  };
-  module.exports.getCouchDbVersion = function(cb) {
-    db.request({}, function(err, body) {
-      if (err) {
-        return cb(err);
-      }
-      var semvers = body.version && body.version.match(/(\d+)\.(\d+)\.(\d+)/);
-
-      cb(null, {
-        major: semvers[1],
-        minor: semvers[2],
-        patch: semvers[3],
-      });
-    });
-  };
 } else {
   logger.info(
     'Please define a COUCH_URL in your environment.\n' +
